+++ tools/device.h
@@ -12,7 +12,7 @@
 
 extern int device_sector_size;
 
-extern int device_open(char *device, int flags);
+extern long device_open(char *device, int flags);
 extern void device_close(void *data);
 extern int device_write_sector(void *data,off_t offset, void* buffer, size_t size);
 extern int device_read_sector(void *data, off_t offset, void* buffer, size_t size);
